add global variable and rename component
