+++ src/App.scss
@@ -9,8 +9,8 @@
     flex-direction: column;
     justify-content: center;
     align-items: center;
+    padding: 60px 0px;
     width: 100%;
-    height: 95vh;
 }
 
 .btn-form {
