+++ src/components/routes/PublicRoute.js
@@ -1,15 +1,15 @@
 import React from 'react';
 import { Route, Redirect } from 'react-router-dom';
 import { isLogin } from '../../Base';
-import NonAuthLayout from "../layouts/NonAuthLayout";
+import PublicLayout from "../layouts/PublicLayout";
 
 const PublicRoute = ({ component: Component, ...rest }) => {
     return (
         <Route {...rest} render={props => (
             !isLogin() ?
-                <NonAuthLayout>
+                <PublicLayout>
                     <Component {...props} />
-                </NonAuthLayout> : <Redirect to="/home" />
+                </PublicLayout> : <Redirect to="/home" />
         )} />
     );
 };
